Enhanced aprgcommon module through refactoring

+++ AprgCommon/AprgCommon/src/Common/Math/Helpers/FractionHelpers.cpp
@@ -74,14 +74,12 @@ void changeFractionToSimplestForm(NumberType1 & numerator, NumberType2 & denomin
 }
 
 // instantiation should fit on long long int
-template void changeFractionToSimplestForm<int, unsigned int, long long int>(int & numerator, unsigned int & denominator);
+template void changeFractionToSimplestForm<int32_t, uint32_t, int64_t>(int32_t & numerator, uint32_t & denominator);
 
 FractionDetails getBestFractionDetailsForDoubleValue(
-        double const doubleValue)
-{
+        double const doubleValue){
     constexpr double tolerance(1E-12);
     FractionDetails result{1, 0, 1};
-
     UnsignedIntegers partialNumerators;
     double fractionalPart(getAbsoluteValue(doubleValue)), doubleValueForNextIteration(doubleValue);
     bool isBeyondUnsignedIntegerLimits(false);
